replace rtrn with global indicating RETURN. keep rtrn as deprecated var for now. call reserCapture by default

+++ lib/test_utils.sh
@@ -29,29 +29,40 @@ tearDown()
 
 capture()
 {
+  resetCapture
+
   $@ >${STD_OUT} 2>${STD_ERR}
-  rtrn=$?
+  RETURN=$?
+  rtrn=${RETURN} # deprecated
 }
 
 resetCapture()
 {
-  rm ${STD_OUT}
-  rm ${STD_ERR}
+  if [ -f ${STD_OUT} ]; then
+    rm ${STD_OUT}
+  fi
+
+  if [ -f ${STD_ERR} ]; then
+    rm ${STD_ERR}
+  fi
+
+  unset RETURN
+  unset rtrn # deprecated
 }
 
 _assertContains()
 {
   if [ 5 -eq $# ]; then
-    msg=$1
+    local msg=$1
     shift
   elif [ ! 4 -eq $# ]; then
     fail "Expected 4 or 5 parameters; Receieved $# parameters"
   fi
 
-  needle=$1
-  haystack=$2
-  expectation=$3
-  haystack_type=$4
+  local needle=$1
+  local haystack=$2
+  local expectation=$3
+  local haystack_type=$4
   
   case "${haystack_type}" in
     "file") grep -q -F -e "${needle}" ${haystack} ;;
@@ -60,8 +71,8 @@ _assertContains()
 
   if [ "${expectation}" != "$?" ]; then
     case "${expectation}" in
-      0) default_msg="Expected <${haystack}> to contain <${needle}>" ;;
-      1) default_msg="Did not expect <${haystack}> to contain <${needle}>" ;;
+      0) local default_msg="Expected <${haystack}> to contain <${needle}>" ;;
+      1) local default_msg="Did not expect <${haystack}> to contain <${needle}>" ;;
     esac
 
     fail "${msg:-${default_msg}}"
